// packages/db/migrations/snapshots/0abbafb354b90e5e9b29c590f3778ef25b78db4ef86080318609eb901022eeaa/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Attachment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly fileURL: CodecTypes['pg/text@1']['output'];
      readonly fileName: CodecTypes['pg/text@1']['output'] | null;
      readonly taskId: CodecTypes['pg/text@1']['output'];
      readonly uploadedById: CodecTypes['pg/text@1']['output'];
    };
    readonly Board: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly projectId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly taskId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Org: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly OrgMember: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly orgId: CodecTypes['pg/text@1']['output'];
      readonly role: 'ADMIN' | 'MEMBER';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly orgId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly ProjectMember: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly projectId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Section: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly status: 'UPCOMING' | 'IN_PROGRESS' | 'DONE';
      readonly boardId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Task: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly sectionId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly priority: 'LOW' | 'MEDIUM' | 'HIGH' | 'URGENT';
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly boardId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly TaskMapping: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly taskId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly profilePictureUrl: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Attachment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly fileURL: CodecTypes['pg/text@1']['input'];
      readonly fileName: CodecTypes['pg/text@1']['input'] | null;
      readonly taskId: CodecTypes['pg/text@1']['input'];
      readonly uploadedById: CodecTypes['pg/text@1']['input'];
    };
    readonly Board: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly projectId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly taskId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Org: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly OrgMember: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly orgId: CodecTypes['pg/text@1']['input'];
      readonly role: 'ADMIN' | 'MEMBER';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly orgId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly ProjectMember: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly projectId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Section: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly status: 'UPCOMING' | 'IN_PROGRESS' | 'DONE';
      readonly boardId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Task: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly sectionId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly priority: 'LOW' | 'MEDIUM' | 'HIGH' | 'URGENT';
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly boardId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly TaskMapping: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly taskId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly profilePictureUrl: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;